reorg struct_pack benchmark
